# proyecto-m2/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bitácora API",
    "version": "1.0.0",
    "description": "API REST para gestionar autores y publicaciones de un blog. Proyecto Integrador del Módulo 2.\n\nTodos los errores responden con el formato `{ \"error\": \"mensaje\" }`; los errores de validación incluyen además `details`."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "tags": [
    {
      "name": "Sistema"
    },
    {
      "name": "Authors"
    },
    {
      "name": "Posts"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Sistema"
        ],
        "summary": "Estado del servidor y la base de datos",
        "responses": {
          "200": {
            "description": "Servidor y base de datos funcionando"
          },
          "503": {
            "description": "La base de datos no responde"
          }
        }
      }
    },
    "/authors": {
      "get": {
        "tags": [
          "Authors"
        ],
        "summary": "Listar autores",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Busca por nombre (sin distinguir mayúsculas)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de autores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Author"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Authors"
        ],
        "summary": "Crear un autor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Autor creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/authors/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID numérico",
          "schema": {
            "type": "integer",
            "minimum": 1
          }
        }
      ],
      "get": {
        "tags": [
          "Authors"
        ],
        "summary": "Obtener un autor",
        "responses": {
          "200": {
            "description": "Autor encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "put": {
        "tags": [
          "Authors"
        ],
        "summary": "Reemplazar un autor (todos los campos)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autor actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "patch": {
        "tags": [
          "Authors"
        ],
        "summary": "Actualizar parcialmente un autor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autor actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Authors"
        ],
        "summary": "Eliminar un autor (y sus posts)",
        "responses": {
          "204": {
            "description": "Autor eliminado"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/authors/{id}/posts": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID numérico",
          "schema": {
            "type": "integer",
            "minimum": 1
          }
        }
      ],
      "get": {
        "tags": [
          "Authors"
        ],
        "summary": "Obtener un autor junto con sus posts",
        "responses": {
          "200": {
            "description": "Autor y sus posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "author": {
                      "$ref": "#/components/schemas/Author"
                    },
                    "posts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Post"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/posts": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Listar posts (más recientes primero)",
        "parameters": [
          {
            "name": "published",
            "in": "query",
            "required": false,
            "description": "Filtra por estado de publicación",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "author_id",
            "in": "query",
            "required": false,
            "description": "Filtra por autor",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "summary": "Crear un post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/posts/author/{authorId}": {
      "parameters": [
        {
          "name": "authorId",
          "in": "path",
          "required": true,
          "description": "ID del autor",
          "schema": {
            "type": "integer",
            "minimum": 1
          }
        }
      ],
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Listar los posts de un autor",
        "responses": {
          "200": {
            "description": "Posts del autor (puede ser un array vacío)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/posts/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID numérico",
          "schema": {
            "type": "integer",
            "minimum": 1
          }
        }
      ],
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Obtener un post",
        "responses": {
          "200": {
            "description": "Post encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "put": {
        "tags": [
          "Posts"
        ],
        "summary": "Reemplazar un post (todos los campos)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "patch": {
        "tags": [
          "Posts"
        ],
        "summary": "Actualizar parcialmente un post (ej. publicarlo)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "summary": "Eliminar un post",
        "responses": {
          "204": {
            "description": "Post eliminado"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Author": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Melanie Kamann"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "melanie@bitacora.dev"
          },
          "bio": {
            "type": "string",
            "nullable": true,
            "example": "Desarrolladora web."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuthorInput": {
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "example": "Melanie Kamann"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "melanie@bitacora.dev"
          },
          "bio": {
            "type": "string",
            "nullable": true,
            "example": "Desarrolladora web."
          }
        }
      },
      "AuthorPatch": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "bio": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "author_id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Cómo funciona el Event Loop"
          },
          "content": {
            "type": "string"
          },
          "published": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "author": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              }
            }
          }
        }
      },
      "PostInput": {
        "type": "object",
        "required": [
          "author_id",
          "title",
          "content"
        ],
        "properties": {
          "author_id": {
            "type": "integer",
            "minimum": 1,
            "example": 1
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "example": "Mi nuevo post"
          },
          "content": {
            "type": "string",
            "example": "Contenido del post"
          },
          "published": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "PostPatch": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "author_id": {
            "type": "integer",
            "minimum": 1
          },
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "content": {
            "type": "string"
          },
          "published": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "invalid author data"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "must be a valid email"
                }
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Parámetro inválido",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Datos inválidos en el body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "NotFound": {
        "description": "Recurso no encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "El email ya está registrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServerError": {
        "description": "Error interno del servidor",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}